src/email: restructure files
I organized the files according to message content type
e.g. plaintext and html
