+++ src/email/html/index.js
@@ -1,0 +1,6 @@
+const path = require('path');
+const inlineStyles = require('./inlineStyles');
+
+module.exports = {
+  standardTemplate: inlineStyles(path.join(__dirname, 'base-template.html')),
+};
