+++ src/email/plaintext/generateVerifyEmail.test.js
@@ -1,15 +1,15 @@
-const auth = require('../auth');
-const isValidAPI = require('../../jest/isValidAPI');
-const removeTokenFromURL = require('../../jest/removeTokenFromURL');
-const getTokenFromURL = require('../../jest/getTokenFromURL');
+const auth = require('../../auth');
+const isValidAPI = require('../../../jest/isValidAPI');
+const removeTokenFromURL = require('../../../jest/removeTokenFromURL');
+const getTokenFromURL = require('../../../jest/getTokenFromURL');
 
-const { generateVerifyEmailMessage } = require('./email');
+const generateVerifyEmail = require('./generateVerifyEmail');
 
-describe('Email template', () => {
+describe('Plaintext email template', () => {
   it('Links to a valid API endpoint', () => {
     const email = 'test@domain.com';
     const hostDomain = 'auth-now.sh';
-    const { url } = generateVerifyEmailMessage({ email, hostDomain });
+    const { url } = generateVerifyEmail({ email, hostDomain });
     const apiPath = removeTokenFromURL(url);
     expect(isValidAPI(apiPath)).toBeTruthy();
   });
@@ -17,7 +17,7 @@ describe('Email template', () => {
   it('Creates a email with a token that is verifiable', () => {
     const email = 'test@domain.com';
     const hostDomain = 'auth-now.sh';
-    const { url } = generateVerifyEmailMessage({ email, hostDomain });
+    const { url } = generateVerifyEmail({ email, hostDomain });
 
     const payload = auth.verify(getTokenFromURL(url));
     expect(payload.id).toEqual(email);
